+++ changelog.md
@@ -1,7 +1,13 @@
 
 
 # Changelog
-* 2020-04-03
+* 2020-04-07 (v1.7.1)
+  * Fixed an error in the draw.grid(...) function. Non-rastered grids were not properly drawn on zoom and
+    pan comibinations.
+* 2020-04-06
+  * Fixed the radius error in the drawutils.circleHandle function. The radius parameter had no effect and circle
+    handles were always rendered with radius 3.
+* 2020-04-03 (v1.7.0)
   * Fixed broken SVG export (after porting to TS).
   * Fixed broken canvas resizer (after porting to TS).
 * 2020-04-01